(feat) lhremote: add unified meta-package for CLI and MCP
Add unscoped `lhremote` meta-package that composes @lhremote/cli and
@lhremote/mcp into a single install: `npx lhremote <command>` for CLI
use and `npx lhremote mcp` for MCP server on stdio.

Key changes:
- New `packages/lhremote` with unified CLI entrypoint
- Extract shared `runStdioServer()` into `@lhremote/mcp/stdio`
- Reorganize `@lhremote/mcp` exports: root is library, `./stdio` is runner
- Add `lhremote-mcp` bin to `@lhremote/mcp` for standalone use
- Rename `@lhremote/cli` bin to `lhremote-cli` to avoid conflict
- Add `composite: true` to cli/mcp tsconfigs for project references

Closes #52
