+++ packages/lhremote/tsconfig.json
@@ -1,12 +1,12 @@
 {
   "extends": "../../tsconfig.base.json",
   "compilerOptions": {
-    "composite": true,
     "outDir": "dist",
     "rootDir": "src"
   },
   "include": ["src"],
   "references": [
-    { "path": "../core" }
+    { "path": "../cli" },
+    { "path": "../mcp" }
   ]
 }
